Move RPC message types to rpc file
